Split MemDB structs into separate files.
