+++ memdb_iterator.go
@@ -1,0 +1,143 @@
+package db
+
+import (
+	"bytes"
+	"context"
+
+	"github.com/google/btree"
+)
+
+const (
+	// Size of the channel buffer between traversal goroutine and iterator. Using an unbuffered
+	// channel causes two context switches per item sent, while buffering allows more work per
+	// context switch. Tuned with benchmarks.
+	chBufferSize = 64
+)
+
+// memDBIterator is a memDB iterator.
+type memDBIterator struct {
+	ch     <-chan *item
+	cancel context.CancelFunc
+	item   *item
+	start  []byte
+	end    []byte
+}
+
+var _ Iterator = (*memDBIterator)(nil)
+
+// newMemDBIterator creates a new memDBIterator.
+func newMemDBIterator(bt *btree.BTree, start []byte, end []byte, reverse bool) *memDBIterator {
+	ctx, cancel := context.WithCancel(context.Background())
+	ch := make(chan *item, chBufferSize)
+	iter := &memDBIterator{
+		ch:     ch,
+		cancel: cancel,
+		start:  start,
+		end:    end,
+	}
+
+	go func() {
+		// Because we use [start, end) for reverse ranges, while btree uses (start, end], we need
+		// the following variables to handle some reverse iteration conditions ourselves.
+		var (
+			skipEqual     []byte
+			abortLessThan []byte
+		)
+		visitor := func(i btree.Item) bool {
+			item := i.(*item)
+			if skipEqual != nil && bytes.Equal(item.key, skipEqual) {
+				skipEqual = nil
+				return true
+			}
+			if abortLessThan != nil && bytes.Compare(item.key, abortLessThan) == -1 {
+				return false
+			}
+			select {
+			case <-ctx.Done():
+				return false
+			case ch <- item:
+				return true
+			}
+		}
+		switch {
+		case start == nil && end == nil && !reverse:
+			bt.Ascend(visitor)
+		case start == nil && end == nil && reverse:
+			bt.Descend(visitor)
+		case end == nil && !reverse:
+			// must handle this specially, since nil is considered less than anything else
+			bt.AscendGreaterOrEqual(newKey(start), visitor)
+		case !reverse:
+			bt.AscendRange(newKey(start), newKey(end), visitor)
+		case end == nil:
+			// abort after start, since we use [start, end) while btree uses (start, end]
+			abortLessThan = start
+			bt.Descend(visitor)
+		default:
+			// skip end and abort after start, since we use [start, end) while btree uses (start, end]
+			skipEqual = end
+			abortLessThan = start
+			bt.DescendLessOrEqual(newKey(end), visitor)
+		}
+		close(ch)
+	}()
+
+	// prime the iterator with the first value, if any
+	if item, ok := <-ch; ok {
+		iter.item = item
+	}
+
+	return iter
+}
+
+// Close implements Iterator.
+func (i *memDBIterator) Close() {
+	i.cancel()
+	for range i.ch { // drain channel
+	}
+	i.item = nil
+}
+
+// Domain implements Iterator.
+func (i *memDBIterator) Domain() ([]byte, []byte) {
+	return i.start, i.end
+}
+
+// Valid implements Iterator.
+func (i *memDBIterator) Valid() bool {
+	return i.item != nil
+}
+
+// Next implements Iterator.
+func (i *memDBIterator) Next() {
+	item, ok := <-i.ch
+	switch {
+	case ok:
+		i.item = item
+	case i.item == nil:
+		panic("called Next() on invalid iterator")
+	default:
+		i.item = nil
+	}
+}
+
+// Error implements Iterator.
+func (i *memDBIterator) Error() error {
+	return nil // famous last words
+}
+
+// Key implements Iterator.
+func (i *memDBIterator) Key() []byte {
+	if i.item == nil {
+		panic("called Key() on invalid iterator")
+	}
+	return i.item.key
+}
+
+// Value implements Iterator.
+func (i *memDBIterator) Value() []byte {
+	if i.item == nil {
+		panic("called Value() on invalid iterator")
+	}
+	return i.item.value
+}
